added new auth methods
