+++ api/urls.py
@@ -1,24 +1,20 @@
 from django.urls import path
-from api.samples.sample_log import TestLog
-from api.views.auth.login import Login
-from api.views.auth.logout import Logout
-from api.views.auth.signup import Signup
-from api.views.auth.delete_user import DeleteUser
-from api.views.auth.change_password import ChangePasswordView
-# from api.views.patients_management.patients_list import PatientsList
+from api.views.old_auth.login import Login
+from api.views.old_auth.logout import Logout
+from api.views.old_auth.signup import Signup
+from api.views.old_auth.delete_user import DeleteUser
+from api.views.old_auth.change_password import ChangePasswordView
 from api.views.patients_management.personal_info import PersonalInfoView,PersonalInfoModifyView,PersonalInfoAllUsersView
-from api.views.patients_management.patients_api import PatientsList
-from api.views.patients_management.patients_api import PatientsDetail
-from api.views.auth.welcome_auth import WelcomeAuth
+from api.views.patients.patients_api import PatientsList
+from api.views.patients.patients_api import PatientsDetail
+from api.views.old_auth.welcome_auth import WelcomeAuth
 from rest_framework.authtoken.views import obtain_auth_token
-
+from rest_framework.authtoken import views
+from api.views.auth.logout import Logout as NewLogout
 
 
 urlpatterns = [
-    #path('', views.index, name='index'),
-    #path('personalInformations/<int:id>/', PersonalInformationsView.as_view()),
-    #path('personalInformations/modify/<int:id>/', PersonalInformationsModifyView.as_view())
-    
+
     # auth view folder 
     path('login/', Login.as_view(), name='api_token_auth'),  # <-- And here
     path('logout/', Logout.as_view(), name="logout"),
@@ -31,6 +27,11 @@ urlpatterns = [
     path('personalInfo/<int:id>/', PersonalInfoView.as_view()),
     path('personalInfo/modify/<int:id>/', PersonalInfoModifyView.as_view()),
     path('personalInfo/', PersonalInfoAllUsersView.as_view()),  # for multiple operations (update all patients)
+
+
+    # New auth management
+    path('auth/api-token-auth/', views.obtain_auth_token),
+    path('auth/logout/', NewLogout.as_view(), name="auth logout"),
 
     # New patients management 
     path('patients/', PatientsList.as_view(), name="patients"),
