+++ api/views/old_auth/delete_user.py
