+++ api/views/old_auth/login.py
@@ -1,6 +1,3 @@
-# import the logging library
-#import logging
-
 from rest_framework.response import Response
 from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.authtoken.models import Token
@@ -9,7 +6,7 @@ from api.models.patients import Patients
 class Login(ObtainAuthToken):
     
     def post(self, request, *args, **kwargs):
-        #logger = logging.getLogger(__name__)
+        
         serializer = self.serializer_class(data=request.data, context={'request': request})
         serializer.is_valid(raise_exception=True)
         user = serializer.validated_data['user'] # user obj
@@ -18,9 +15,6 @@ class Login(ObtainAuthToken):
         try:
             patient = Patients.objects.get(fk_user_id=user.id) # patient obj
         except:
-            # Get an instance of a logger
-            #logger.error("error catched")
-            print("error catched")
             return Response({"msg":"application error"}, status=500)
                     
         return Response({
