+++ test_with_curl/samples.txt
@@ -1,16 +1,20 @@
 #------ AUTH -----------------
 
-#Signup #Not working
-curl -X POST http://127.0.0.1:8000/api/signup/ \
-     -d 'username="valeriak6@gmail.com"' \
-     -d 'password1="pluto123!"' \
-     -d 'password2="pluto123!"'
-
 # Login 
 curl -X POST http://127.0.0.1:8000/api/login/ \
      -H "Content-Type: application/json" \
-     -d '{"username":"valeriak6@gmail.com","password":"pluto125!"}' 
-     
+     -d '{"username":"login_tester","password":"2MZq5W$rZFDK.y+"}' 
+
+curl -X POST http://127.0.0.1:8000/api/login/ \
+     --form 'username="login_tester"' \
+     --form 'password="2MZq5W$rZFDK.y+"' 
+
+#Signup 
+curl -X POST 'http://127.0.0.1:8000/api/signup/' \
+     --form 'username="login_tester_2"' \
+     --form 'password1="pluto123!"' \
+     --form 'password2="pluto123!"'
+
 #Welcome Auth
 curl -X GET http://127.0.0.1:8000/api/welcomeAuth/ \
      -H 'Authorization: Token f51c12462f5d8899f8441839c564a70d959c4870'
@@ -38,18 +42,30 @@ curl -X GET http://localhost:8000/api/personalInfo/1/
 curl -X POST http://localhost:8000/api/personalInfo/ \
      -H 'Content-Type: application/json' \
 
+
 #------ NEW PATIENT MANAGEMENT -----------------
 
 curl -X GET http://localhost:8000/api/patients/ #Get all patients
 
-curl -X GET http://localhost:8000/api/patients/1/
+curl -X GET http://localhost:8000/api/patients/1/ # Get patient info
 
-curl -X POST http://localhost:8000/api/patients/ \ 
+curl -X POST http://localhost:8000/api/patients/ \ # Add new user
      -H "Content-Type: application/json" \
-     -d '{"firstname": "peppe","lastname": "peppe","birthDate": 1992-04-17,"gender": "M","bloodType": 4,"fk_user": null}'
+     -d '{"id": 1,"firstname": "valeria","lastname": "dolce","birthDate": "1992-04-13","gender": "F"}'
 
 curl -X PUT http://localhost:8000/api/patients/1/ \
      -H "Content-Type: application/json" \
      -d '{"id": 1,"firstname": "valeria","lastname": "dolce","birthDate": "1992-04-13","gender": "F","bloodType": 4,"fk_user": 3}'
 
 curl -X DELETE http://localhost:8000/api/patients/1/
+
+#------ NEW AUTH MANAGEMENT -----------------
+
+#Login
+curl -X POST http://127.0.0.1:8000/api/api-token-auth/ \
+     -H "Content-Type: application/json" \
+     -d '{"username":"login_tester","password":"2MZq5W$rZFDK.y+"}' 
+
+#Logout
+curl -X GET http://127.0.0.1:8000/api/auth/logout/ \
+     -H 'Authorization: Token 3ad451a132d225d54f8963652866e01d388649cb'
